# src/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "my-store",
    "description": "Swagger documentation for my-store..."
  },
  "host": "localhost:3000",
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/api/user/register": {
      "post": {
        "tags": [
          "Register user"
        ],
        "summary": "Returns a user by id",
        "description": "This endpoint will return a user by id...",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/api/user/login": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/user/profile": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "version",
            "in": "header",
            "type": "string"
          },
          {
            "name": "filter",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "someBody": {
        "$name": "Jhon Doe",
        "$age": 29,
        "about": ""
      },
      "someResponse": {
        "name": "Jhon Doe",
        "age": 29,
        "diplomas": [
          {
            "school": "XYZ University",
            "year": 2020,
            "completed": true,
            "internship": {
              "hours": 290,
              "location": "XYZ Company"
            }
          }
        ]
      },
      "someEnum": {
        "@enum": [
          "red",
          "yellow",
          "green"
        ]
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}